do not set expiry date, if the "expiration" setting is 0
reason: for id tags without expiry dates, we until now always have set an
expiry date as "now + {hoursToExpire}" (to force the station ask steve
as frequent as possible about auth info). this might not always be the
desired behavior. for people who want to completely disable this, make it
possible to do so.